#244 feat: WIP - Update UI
