+++ applications/portal/frontend/package.json
@@ -112,5 +112,8 @@
     "webpack-cli": "^4.6.0",
     "webpack-dev-server": "^3.11.2",
     "webpack-merge": "^5.7.0"
+  },
+  "resolutions": {
+    "@types/react": "17.0.38"
   }
 }
